Refactor live/replay representation (#245)
While working on adding a replay overlay (#240), I found it weird that we have to dig so deeply to extract the `LiveOrReplay` (formerly `MidRoundStateVariant`; see #247). In particular, doing so requires one `case` branch for `Spawning` and one for `Moving`, which feels backwards to me.

This PR therefore moves `LiveOrReplay` to the same "level" as `PausedOrNot`, which I think feels more natural. Here are a couple of observations that I think support my reasoning:

  * `reactToTick` and `consumeAnimationFrame` can no longer swap out `Live` for `Replay` or vice versa.
  * `playOutRound` in the `TestHelpers` module no longer needs to arbitrarily pick either `Live` or `Replay`; it's instead just completely agnostic to that.

And a summary of the concrete changes:

  * The `MidRoundState` type is replaced with `Round` everywhere it's used.
  * `modifyRound` is removed because it would otherwise just be the identity function.
  * A `LiveOrReplay` is added everywhere that information is needed.

Variable names like `midRoundState` are kept at various locations, because they still refer to a mid-round state, albeit not a `MidRoundState`.

💡 `git show --color-words=' _,|\w+|.'`