Add replay indicator (#240)
The replay feature introduced in a9eec802731850406b8b9d7c0fbf64d4e039b201 is great, but its UX isn't. What usually happens when replaying a round is that players think it's a live round, causing these problems:

  * Players who aren't ready for the next round are taken by surprise and scramble to get their hands on the keyboard in panic.
  * Players who are ready are confused because they cannot control their Kurve.

This PR should make it easier to tell the difference between live and replay, by showing a vintage "R" replay indicator/overlay in the top-left corner. It's transparent so as not to obscure something that may have happened right behind it.

The replay indicator is _not_ shown when a replayed round is over. Whether we would like to show it or not, we currently can't, because the `RoundOver` variant doesn't contain a `LiveOrReplay`. Also, @lydell [expressed] this opinion:

> I actually like it when the “R” disappears when the replay is over and you’re back in “round is over, time to press Space to continue” mode.

These PRs laid the groundwork for this one:

  * #242
  * #244
  * #247
  * #245
  * #249

[expressed]: https://github.com/SimonAlling/kurve/pull/240#discussion_r2617096784

💡 `git show --color-words=.`