Enable pausing a replay (#250)
When replaying a round, it's useful to be able to pause, for example to explain to other players where they should be looking, what's going to happen, etc. It seems reasonable to use the Space key for that, analogously to how media players typically work.

Unlike when a replay is paused due to loss of focus, when pausing manually using Space, the "Press Space to continue" hint is really annoying and only states what's already obvious.

However, the model doesn't contain enough information to selectively show the hint when the replay was paused due to loss of focus, but not when it was paused manually. Therefore, I'm seizing the opportunity to also make loss of focus _not_ auto-pause a replay, because the entire point of auto-pausing is to minimize the damage caused by focus loss in a _live_ round.

It might seem like this PR should be split into two, but it's actually quite reasonable to make those changes together, because then we continuously stay at exactly one way to end up in an `Active Replay Paused _` game state. Consider the alternatives:

  * **Remove auto-pausing on focus loss first.** That would make the `Active Replay Paused _` branch in `TextOverlay.content` unreachable (unbeknownst to the type checker).
  * **Add manual pausing first.** That would make the `Active Replay Paused _` branch difficult or impossible to implement "correctly", because there would be two distinctly different ways to reach it, each with its own desired implementation.

Making both changes together means that we can just change the `Active Replay Paused _` branch from having the hint to not having it.

As it happens, the default case in `update` doesn't type check anymore, because it's trivially unreachable. It's therefore removed.

💡 `git show --color-words='\w+|.'`