Enable skipping ahead in a replay (#252)
It can be extremely boring to sit around waiting for a replay to play out in real time. This PR makes it possible to fast-forward 5 seconds at a time using the right arrow key, similarly to how media players typically work.

The implementation is literally just the `AnimationFrame` branch with `delta` replaced by 5000 milliseconds.

## Limitations

  * Going _back_ is much harder to implement, so that's not included here.
  * Skipping during spawning isn't enabled here, but hopefully soon will be.