Make reactToTick Cmd-free (#254)
I don't like that we have so much `Cmd` sprinkled across the code base. `Cmd`-returning functions are harder to test, the order of batched commands is not specified, etc. This PR makes `reactToTick` return a description of what do draw, which we can then derive a `Cmd` from. That is, `reactToTick` doesn't use `headDrawingCmd` or `bodyDrawingCmd` anymore, and they are therefore unexposed in the `Canvas` module.

💡 `git show --color-words='drawingCmd|[A-Z]?[a-z]+|.'`