Fix broken head rendering (#257)
## Problem

Turns out that #255, which fixed a rendering bug, introduced another one instead: heads appear transparent and/or flickering in the middle of holes, and they may be completely invisible when a round is paused in the middle of a hole.

This change makes it glaringly obvious, especially on a 240 Hz monitor:

```diff
--- a/src/Config.elm
+++ b/src/Config.elm
@@ -26,6 +26,6 @@ default =
         , holes =
-            { minInterval = Distance 90
-            , maxInterval = Distance 300
-            , minSize = Distance 5
-            , maxSize = Distance 9
+            { minInterval = Distance 50
+            , maxInterval = Distance 50
+            , minSize = Distance 500
+            , maxSize = Distance 500
             }
```

The reason is that `Cmd.none` was replaced with `nothingToDraw` in `consumeAnimationFrame`. That may sound equivalent, but it's not:

  * `Cmd.none` means "do absolutely nothing at all".
  * `nothingToDraw` means "call the `renderOverlay` port with the empty list and call the `render` port with the empty list".

When the `renderOverlay` port is called, it always clears the entire overlay canvas before drawing anything. If it's told to draw the empty list, then the result is that the overlay canvas is cleared.

The problem manifests itself whenever an animation frame doesn't consume at least one tick, i.e. whenever `recurse` immediately ends up in the `else` branch without calling itself recursively. This happens on roughly one in two animation frames at 120 Hz, roughly one in four animation frames at 240 Hz, etc.

## Solution

This PR fixes the problem by making "do absolutely nothing at all" representable. To achieve that in a clean way, `WhatToDraw` is replaced with a more expressive type, `RenderAction`, whose `LeaveAsIs` variant corresponds to `Cmd.none`.

💡 `git show --color-words='.+(head|body)Drawing = \[\]|whatToDraw|[A-Z]?[a-z]+|.'`