Refactor spawn rendering (#258)
Today, the flickering when Kurves are spawning is implemented by alternating between these commands:

  1. Drawing a _body_ square in the spawning Kurve's color
  2. Clearing that square

This PR replaces that with alternating between these commands:

  1. Drawing the _head_ of the spawning Kurve _and_ the already spawned ones
  2. Drawing only the heads of the already spawned ones

The change from "body" to "head" is significant: heads are drawn on the overlay canvas, which is always cleared before anything is drawn on it. Therefore, the already spawned Kurves need to be explicitly drawn in every spawn tick.

The end goal is to simplify rendering overall; I think this takes us one step closer.

## `drawSpawnsPermanently`

Today, we rely on `drawSpawnIfAndOnlyIf` being called with `True` when `ticksLeft` is even (because the final value of `ticksLeft` is `0`). Inverting that logic would make the spawn square seemingly disappear when the Kurve starts moving (but its hitbox would still be there). Here's the easiest way to demonstrate that (without the changes in this PR, of course):

```diff
--- a/src/Main.elm
+++ b/src/Main.elm
@@ -125,7 +125,7 @@ stepSpawnState config { kurvesLeft, ticksLeft } =
                     else
                         { kurvesLeft = spawning :: waiting, ticksLeft = ticksLeft - 1 }
             in
-            ( Just newSpawnState, drawSpawnIfAndOnlyIf (isEven ticksLeft) spawning )
+            ( Just newSpawnState, drawSpawnIfAndOnlyIf (not <| isEven ticksLeft) spawning )


 update : Msg -> Model -> ( Model, Cmd Msg )
```

With the flickering heads now being drawn on the auto-cleared overlay canvas instead of the persistent main canvas, body squares must be explicitly "committed" to the main canvas when spawning is complete and the Kurves are about to start moving. `drawSpawnsPermanently` does that.

## Further simplification

With this PR, the `clear` port is only used for one thing (clearing the entire main canvas), which will let us simplify/specialize it. However, that's not done here, in order to keep diff size and complexity down, since changes concerning `Cmd`s tend to be the most tricky ones and introduce the most surprising problems.

💡 `git show --color-words='alreadySpawnedKurves |headDrawingCmd .+|\w+|.'`