Remove test case for cutting corners even more (#260)
The mechanic covered by the test case added in #125 is covered much better by the test case added in #126, which is easy to reproduce in the original game (since #197) thanks to the insights gained in #93. In contrast, the exact positions and angles needed to reproduce the test case removed in this PR are not known.

Note that the screenshot added in #125 was incorrectly named, and therefore subsequently renamed in #127.