Use .5 coordinates consistently in test cases (#261)
We've usually found test-case scenarios with positions like (100.5, 60.5) instead of (100, 60) easier to reason about, because (100.5, 60.5) is the exact center of the pixelated square representing that position on the canvas. This PR makes all test-case scenarios use such coordinates.

The only test-case scenarios visually and/or semantically affected by this PR are these:

  * `CrashIntoKurveTiming`
  * `CrashIntoWallExactTiming`

Green's "upward snapping" is delayed by one tick, causing Green to die one tick later.

## Checking if test cases are visually affected

I wrote this script:

```bash
#!/bin/bash

set -euo pipefail

visualizeTestScenario() {
  scenarioName="$1"
  git restore src/Main.elm
  git apply <<EOF
--- a/src/Main.elm
+++ b/src/Main.elm
@@ -47,2 +47,3 @@ import Round exposing (Round, initialStateForReplaying, modifyAlive, modifyKurve
 import Set exposing (Set)
+import TestScenarios.${scenarioName}
 import Time
@@ -66,4 +67,12 @@ init : () -> ( Model, Cmd Msg )
 init _ =
+    let
+        ( gameState, cmd ) =
+            { seedAfterSpawn = Random.initialSeed 0
+            , spawnedKurves = TestScenarios.${scenarioName}.spawnedKurves
+            }
+                |> prepareReplayRound
+                |> newRoundGameStateAndCmd Config.default Replay
+    in
     ( { pressedButtons = Set.empty
-      , appState = InMenu SplashScreen (Random.initialSeed 1337)
+      , appState = InGame gameState
       , config = Config.default
@@ -71,3 +80,3 @@ init _ =
       }
-    , Cmd.none
+    , cmd
     )
EOF
}

for f in src/TestScenarios/*.elm; do
  scenarioName=$(basename --suffix=".elm" $f)
  visualizeTestScenario "$scenarioName"
  { while ! npm run build; do sleep 2; done }
  echo -n "💡 CURRENT: ${scenarioName} "
  read -p "(Enter to continue)"
done
```

Then I ran `npm start` in two [worktrees]: without and with the changes in this PR, respectively. In each worktree, I used the script above to go through the scenarios one by one. In the meantime, I had http://localhost:8000 (without this PR) and http://localhost:8001 (with this PR) open in adjacent browser tabs, which I switched between to check for visual changes.

I ran into these problems:

  * I had to delete `npm run install-hooks && ` in the alternative worktree for `npm start` to work:

    ```diff
    --- a/package.json
    +++ b/package.json
    @@ -12,3 +12,3 @@
        "test": "elm-test",
    -    "start": "npm run install-hooks && run-pty % elm-watch hot % esbuild --serve --servedir=./_site/ % sass --watch --source-map src/css/:_site/css/"
    +    "start": "run-pty % elm-watch hot % esbuild --serve --servedir=./_site/ % sass --watch --source-map src/css/:_site/css/"
      },
    ```

  * The `CrashIntoKurveTiming` scenario required manual modifications in `Main.elm`, because unlike the others, its `spawnedKurves` takes an argument. The effect of this PR on the visualization was the same for all these arguments:

      * `100 + toFloat 0 / 10`
      * `100 + toFloat 5 / 10`
      * `100 + toFloat 9 / 10`

I don't know if `100` should be changed to `100.5` in the definition of `y_red`, but I think it makes sense to keep it, so that we create test cases for 100.0, 100.1, …, 100.9.

[worktrees]: https://git-scm.com/docs/git-worktree

💡 `git show --color-words='[0-9.]+|.'`