Align position representation with original game (#262)
Our understanding of the original game from #93 is that internal positions concern the _top left corner_ of the Kurve's 3 × 3 head, not its center. For example, the canonical way to represent a Kurve's head being in the very top left corner of the canvas should perhaps be thought of as (0, 0), not (1.5, 1.5). Furthermore, the floating-point components of the internal position are not rounded to the nearest integer: for example, 5.9 is rounded to 5, not 6.

This PR makes our representation match that of the original game better, which should make it easier to compare scenarios between the original game and our clone. Specifically, `edgeOfSquare` is modified to 1) not subtract 1.5 anymore, and 2) use `floor` instead of `round`.

To preserve the visual appearance and semantics of test-case scenarios, all internal positions are modified to map to the same _drawing_ positions, by subtracting 1 from each position component. It can be thought of as the two changes to `edgeOfSquare` partially canceling each other out, leaving a difference of 1 that must be compensated for. I used the same technique as in #261 to check if test-case scenarios were visually affected.

## Future improvements

Our current understanding is that the original game actually rounds _towards zero_, not down – that is, we should use `truncate` instead of `floor`. That's deliberately left out here though, because it's so incredibly counterintuitive that it deserves its own PR.

Resolves #191.

💡 `git show --color-words='round|[0-9]+|.'`

Co-authored-by: Simon Lydell <simon.lydell@gmail.com>