Replicate off-by-one error at top and left wall (#263)
As briefly mentioned in #262, our understanding is that the original game derives drawing positions not by rounding down, but by _truncating_ – i.e. effectively just removing the decimal parts of the floating-point components of the internal position. That results in counterintuitive behavior at the top and left walls, as made obvious by this scenario:

```elm
theScenario : Scenario
theScenario =
    [ ( Red
      , { x = 50
        , y = 3
        , direction = pi / 2 + 0.01
        }
      )
    , ( Green
      , { x = 50
        , y = 100
        , direction = pi / 2
        }
      )
    ]
```

Note that Red moves ~100 pixels parallel to the wall, then "snaps" up 1 pixel, then moves another ~100 pixels, then "snaps" up again – and then moves about _200 pixels_ before crashing. Without deep familiarity with the original game, one would expect it to crash much sooner.

For illustrative purposes, consider the above scenario, but with `0.2` instead of `0.01`. You'd be forgiven for expecting it to end something like this:

    ⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛🟥🟥🟥🟥🟥🟥🟥⬛⬛⬛⬛⬛⬛⬛
    ⬛⬛⬛⬛⬛⬛⬛🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥⬛⬛⬛⬛⬛⬛⬛
    ⬛⬛🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥⬛⬛⬛⬛⬛⬛⬛
    ⬛⬛🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛
    ⬛⬛🟥🟥🟥🟥🟥🟥🟥⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛
    ⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛
    ⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛

But it ends like this:

    ⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥⬛⬛
    ⬛⬛⬛⬛⬛⬛⬛🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥⬛⬛
    ⬛⬛🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥⬛⬛
    ⬛⬛🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥🟥⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛
    ⬛⬛🟥🟥🟥🟥🟥🟥🟥⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛
    ⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛
    ⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛⬛

The reason is that coordinates in the interval (0, –1) are rounded to 0 for hit detection and drawing. So it's effectively possible for a Kurve to be slightly outside the top and left boundaries of the canvas, but that's _rendered_ as if it's completely within the canvas.

💡 `git show --color-words='floor|[0-9]+|.'`

Co-authored-by: Simon Lydell <simon.lydell@gmail.com>