Replace RenderAction with WhatToDraw where possible (#273)
Since #270, `drawSpawnsPermanently` and `drawSpawnIfAndOnlyIf` return a `RenderAction`, even though they can never return `LeaveAsIs`. That is, they always just return information equivalent to a `WhatToDraw`, but we hide that knowledge from the type checker. This PR changes them to returning a `WhatToDraw`.

In other words, this PR does for `drawSpawnsPermanently` and `drawSpawnIfAndOnlyIf` what #271 did for `reactToTick`.

💡 `git show --color-words='\w+|.'`