Replace RenderAction with WhatToDraw even more (#274)
Today, `mergeRenderAction` takes two `RenderAction`s and returns a `RenderAction`, even though it's never called with `LeaveAsIs` as the second argument and therefore will never return `LeaveAsIs`. That is, we always pass information equivalent to a `WhatToDraw` as the second argument, and always receive such information back, but we hide that knowledge from the type checker. This PR changes the function to taking only one `RenderAction` (not two) and returning a `WhatToDraw`, and renames it accordingly.

In other words, this PR does for `mergeRenderAction` approximately what #271 did for `reactToTick` and #273 did for `drawSpawnsPermanently` and `drawSpawnIfAndOnlyIf`.

💡 `git show --color-words='RenderAction|action|LeaveAsIs|\( .|.'`