Use Effect pattern (#279)
This PR builds further upon #254, #255 and #270 by making `update` not return a `Cmd` anymore, but instead an `Effect` – a transparent description of what to do. This is commonly referred to as "the Effect pattern" or ["the effects pattern"][effects].

This should make `update` much easier to write tests for, and perhaps reason about. Instead of constructing `Cmd`s all over the place, we only do that using `makeCmd` in `updateWithCmd`, which is just a thin wrapper around `update`.

Here's a summary of how the new representation corresponds to the old one:

| Old               | New               |
|-------------------|-------------------|
| `Cmd msg`         | `Effect`          |
| `drawingCmd`      | `DrawSomething`   |
| `clearEverything` | `ClearEverything` |
| `Cmd.none`        | `DoNothing`       |

[effects]: https://sporto.github.io/elm-patterns/architecture/effects.html

💡 `git show --color-words='[A-Z]?[a-z]+|.'`

Co-authored-by: Simon Lydell <simon.lydell@gmail.com>