Refactor tick logic slightly in playOutRound (#286)
I'm fighting a never-ending struggle to wrap my head around the _meaning_ of a tick: what it means to "be at" a tick or "perform" a tick, what the "current" tick means, etc. I feel like I'm walking through a minefield of off-by-one errors every time.

This PR makes these renamings in `playOutRound`:

  * `tick` → `lastTick`
  * `nextTick` → `incrementedTick`

That makes `recurse` in `playOutRound` more similar to its namesake in `consumeAnimationFrame`.

This PR also reuses the incremented tick in the base case, instead of computing it again.

💡 `git show --color-words='\w+|.'`