Hardcode game-over logic (#288)
Ever since it became possible for the game to end in #59, it has been defined in the `Config` _when_ it ends. However, it's defined as a function, which is potentially problematic in the context of end-to-end testing. If we're going to expect an initial model and a sequence of messages to result in some model, then if the model contains a function, we cannot describe the expected model, at least not without having access to the exact function.

This PR hardcodes the game-over logic. We might still want to make it more customizable in the future, but until then: [YAGNI].

[YAGNI]: https://en.wikipedia.org/wiki/You_aren%27t_gonna_need_it